requirements.txt
crewai==0.86.0
crewai-tools==0.17.0
langchain==0.3.11
langchain-community==0.3.11
langchain-text-splitters==0.3.2
chromadb==0.5.23
mem0ai==0.1.34
groq==0.13.0
sentence-transformers==3.3.1
fastapi==0.115.6
uvicorn==0.32.1
python-multipart==0.0.20
streamlit==1.41.1
requests==2.32.3
passlib==1.7.4